Move integration test utils into new module test.on_yubikey.framework
