+++ test/on_yubikey/cli_piv/test_management_key.py
@@ -1,7 +1,7 @@
 import re
 import unittest
 
-from ..util import cli_test_suite
+from ..framework import cli_test_suite
 from .util import (
     old_new_new, DEFAULT_PIN, DEFAULT_MANAGEMENT_KEY,
     NON_DEFAULT_MANAGEMENT_KEY)
