+++ test/on_yubikey/test_cli_misc.py
@@ -1,7 +1,6 @@
 import unittest
 
-from . import yubikey_conditions
-from .util import (cli_test_suite)
+from .framework import cli_test_suite, yubikey_conditions
 
 
 @cli_test_suite
